# openapi.yaml
openapi: 3.0.3
info:
  title: Vision Financial API
  version: "1.0"
  description: >
    Self-hosted financial transaction manager.
    All responses are wrapped in the ADR-026 envelope:
    `{ ok: true, data: T }` on success, `{ ok: false, error: {...} }` on failure.
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Shared envelope & list wrappers ─────────────────────────────────────────
components:
  parameters:
...
    OptionalLimit:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
      description: >
        Page size. Omit (together with offset) to receive the whole collection.
        Values above the endpoint's cap are clamped.
    OptionalOffset:
      in: query
      name: offset
      required: false
      schema:
        type: integer
        minimum: 0
      description: Rows to skip. Omit (together with limit) to receive the whole collection.
  schemas:
...
    Envelope:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        data:
          description: Present on success
        error:
          $ref: "#/components/schemas/ErrorPayload"
...
    CategoryCreate:
      type: object
      required: [general, detail]
      properties:
        general:
          type: string
        detail:
          type: string
        description:
          type: string
...
    CategoryList:
      allOf:
        - $ref: "#/components/schemas/OptionalPaginationFields"
        - type: object
          required: [items, links]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Category"
            links:
              type: array
              items:
                $ref: "#/components/schemas/Link"

    # ─── Recipients ─────────────────────────────────────────────────────────
...
    RecipientCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        default_category_id:
          type: integer
        notes:
          type: string
...
paths:
...
  /api/categories:
    get:
      summary: List categories
      operationId: getCategories
      tags: [Categories]
      parameters:
        - $ref: "#/components/parameters/OptionalLimit"
        - $ref: "#/components/parameters/OptionalOffset"
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: is_active
          schema:
            type: boolean
      responses:
        "200":
          description: Category list (complete unless limit/offset are supplied)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CategoryList"
    post:
      summary: Create or get existing category
      operationId: createCategory
      tags: [Categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
      responses:
        "200":
          description: Existing category returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Category"
        "201":
          description: New category created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Category"
...
  /api/recipients:
    get:
      summary: List recipients
      operationId: getRecipients
      tags: [Recipients]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: is_active
          schema:
            type: boolean
      responses:
        "200":
          description: Recipient list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecipientList"
    post:
      summary: Create or get existing recipient
      operationId: createRecipient
      tags: [Recipients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipientCreate"
      responses:
        "200":
          description: Existing recipient
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Recipient"
        "201":
          description: New recipient created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Recipient"
...
  /api/import/csv:
    post:
      summary: Import transactions from bank CSV
      operationId: importCsv
      tags: [Import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, bank_name]
              properties:
                file:
                  type: string
                  format: binary
                bank_name:
                  type: string
                  description: Bank adapter identifier. The route also accepts it
                    as a query parameter (`req.query.bank_name || req.body.bank_name`,
                    routes/importRoutes.js) — the frontend client sends it in the
                    query string.
      responses:
        "201":
          description: All rows matched exactly — batch committed immediately
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImportCsvResult"
        "202":
          description: Some rows need review — batch left in awaiting_review; client must use the review endpoints then POST commit
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImportCsvReviewRequired"
  /api/import/csv/custom:
    post:
      summary: Import CSV with custom column mapping
      operationId: importCsvCustom
      tags: [Import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Flat form fields (customCsvImportSchema in
                routes/importRoutes.js). The route merges query and body
                (`{...req.query, ...req.body}`), so every field may equally be
                sent as a query parameter — the frontend client sends them all
                in the query string.
              required: [file, bank_name, date_format, date_column, recipient_column, amount_column]
              properties:
                file:
                  type: string
                  format: binary
                bank_name:
                  type: string
                  description: Display label for the custom adapter
                date_format:
                  type: string
                  description: Date format of the date column (e.g. DD/MM/YYYY)
                date_column:
                  type: string
                recipient_column:
                  type: string
                amount_column:
                  type: string
                memo_column:
                  type: string
                separator:
                  type: string
                  description: Single-character CSV delimiter, default ','
                encoding:
                  type: string
                  default: utf-8
                skip_rows:
                  type: integer
                  default: 0
                  minimum: 0
      responses:
        "201":
          description: All rows matched exactly — batch committed immediately
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImportCsvResult"
        "202":
          description: Some rows need review — batch left in awaiting_review; client must use the review endpoints then POST commit
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImportCsvReviewRequired"
  /api/import/csv/stream:
    post:
      summary: Stream-import large CSV file
      operationId: importCsvStream
      tags: [Import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, bank_name]
              properties:
                file:
                  type: string
                  format: binary
                bank_name:
                  type: string
      responses:
        "200":
          description: Server-Sent Events stream; events are progress /
            review_required / complete / error. The `complete` event carries
            `{total_processed, imported, duplicates, errors, batch_id,
            auto_linked_count, status, percent}` (routes/importRoutes.js
            buildComplete); `review_required` carries `{batch_id,
            match_source_counts, percent}` — no counts. Event payloads are not
            modelled as schemas (OpenAPI cannot type SSE events), so the
            frontend keeps a hand-written `ImportResult` type for this stream.
          content:
            text/event-stream:
              schema:
                type: string
  /api/import/recipients:
    post:
      summary: Bulk-import recipients from CSV
      operationId: importRecipients
      tags: [Import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                separator:
                  type: string
                  minLength: 1
                  maxLength: 1
                  default: ","
                  description: Single-character CSV delimiter.
                encoding:
                  type: string
                  default: utf-8
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/import/categories:
    post:
      summary: Bulk-import categories from CSV
      operationId: importCategories
      tags: [Import]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                separator:
                  type: string
                  minLength: 1
                  maxLength: 1
                  default: ","
                  description: Single-character CSV delimiter.
                encoding:
                  type: string
                  default: utf-8
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
  /api/import/batches/{id}/preview:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
    get:
      summary: Preview import batch rows before commit
      operationId: previewImportBatch
      tags: [Import]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Row preview with staged data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/import/batches/{id}/rows/{rowId}/override:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
      - in: path
        name: rowId
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
          maximum: 9007199254740991
    post:
      summary: Override the matched recipient for a staged row
      operationId: overrideImportRow
      tags: [Import]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient_id:
                  type: integer
                  format: int32
                  minimum: 1
                  nullable: true
                  description: >-
                    Recipient to attribute the staged row to. `null` — or an
                    absent field — clears the override. A present value must be
                    a positive integer; a malformed one is rejected rather than
                    coerced, because a coerced id names a different recipient
                    and the batch then commits the transaction against it.
      responses:
        "200":
          description: Row updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
        "400":
          description: Malformed batch id, row id or recipient_id
        "404":
          description: Row not found in this batch, or not in matched status
...